requirements/common.txt
Django==1.9.1
bcrypt==2.0.0
psycopg2==2.6.1
mysqlclient==1.3.7
django-phonenumber-field==1.1.0
csvkit==0.9.1
django-cors-headers==0.12
djangorestframework==3.3.2 # REST API
djangorestframework-camel-case==0.2.0
Markdown==2.6.5       # Markdown support for the browsable API.
# django-filter==0.12.0  # Filtering support
django-oauth-toolkit==0.10.0

ipaddress==1.0.16


# Emails
django-premailer==0.1.0

# Tesseract OCR Client
pytesseract==0.1
PyPDF2==1.25.1
Pillow==3.1.1

# GeoIP Service
geoip2==2.2.0

gunicorn==19.4.5
django-celery==3.1.17

django-storages==1.4.1
boto==2.40.0

django-extensions==1.6.1
dj-database-url==0.3.0

requests==2.9.1

raven==5.11.1

ipython==4.0.1
ipdb==0.8.1

factory-boy==2.6.1
mock==2.0.0
funcsigs==1.0.0
nose==1.3.7
flake8==2.5.4
